add: cache service 프로토콜 구현
